// type.ts
 export type Iuser ={
    name: string
    email: string
    password: string
}
export type Body= {
    email: string
    password: string
  }
  export type Address={
    country: String
    city: String
    street: String
    userId :Number
  }
  export type books={
    title: String
  description: String
  }